Fix missing resource breaking old tests
